Refactor Configuration Plugin
