+++ antibytes-configuration/src/main/kotlin/tech/antibytes/gradle/configuration/AndroidConfigurationApiContract.kt
@@ -8,7 +8,7 @@ package tech.antibytes.gradle.configuration
 
 import org.gradle.api.JavaVersion
 
-internal interface ConfigurationApiContract {
+internal interface AndroidConfigurationApiContract {
     interface TestSource {
         val sourceDirectories: Set<String>
         val resourceDirectories: Set<String>
@@ -62,7 +62,6 @@ internal interface ConfigurationApiContract {
     }
 
     companion object {
-        // Android
         const val ANDROID_PREFIX = "antibytes"
         const val ANDROID_PREFIX_SEPARATOR = "_"
         const val TARGET_SDK = 33
@@ -71,8 +70,5 @@ internal interface ConfigurationApiContract {
         val COMPATIBILITY_TARGETS = JavaVersion.VERSION_1_8
         const val TEST_RUNNER = "androidx.test.runner.AndroidJUnitRunner"
         val TEST_RUNNER_ARGUMENTS = mapOf("clearPackageData" to "true")
-
-        // iOS
-        const val IOS_DEFAULT_DEVICE = "iPhone 14"
     }
 }
