+++ antibytes-configuration/src/main/kotlin/tech/antibytes/gradle/configuration/AntiBytesConfiguration.kt
@@ -8,6 +8,10 @@ package tech.antibytes.gradle.configuration
 
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import tech.antibytes.gradle.configuration.android.AndroidApplicationConfigurator
+import tech.antibytes.gradle.configuration.android.AndroidLibraryConfigurator
+import tech.antibytes.gradle.configuration.android.DefaultAndroidApplicationConfigurationProvider
+import tech.antibytes.gradle.configuration.android.DefaultAndroidLibraryConfigurationProvider
 import tech.antibytes.gradle.util.isAndroidApplication
 import tech.antibytes.gradle.util.isAndroidLibrary
