+++ antibytes-configuration/src/main/kotlin/tech/antibytes/gradle/configuration/IosConfigurationApiContract.kt
@@ -1,0 +1,14 @@
+/*
+ * Copyright (c) 2022 Matthias Geisler (bitPogo) / All rights reserved.
+ *
+ * Use of this source code is governed by Apache License, Version 2.0
+ */
+
+package tech.antibytes.gradle.configuration
+
+internal interface IosConfigurationApiContract {
+    companion object {
+        // iOS
+        const val IOS_DEFAULT_DEVICE = "iPhone 14"
+    }
+}
