+++ antibytes-configuration/src/main/kotlin/tech/antibytes/gradle/configuration/android/AndroidApplicationConfigurator.kt
@@ -4,20 +4,19 @@
  * Use of this source code is governed by Apache License, Version 2.0
  */
 
-package tech.antibytes.gradle.configuration
+package tech.antibytes.gradle.configuration.android
 
-import com.android.build.api.dsl.LibraryExtension
+import com.android.build.api.dsl.ApplicationExtension
 import org.gradle.api.Project
-import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
-import tech.antibytes.gradle.util.isKmp
+import tech.antibytes.gradle.configuration.AndroidConfigurationApiContract
+import tech.antibytes.gradle.configuration.ConfigurationContract
 
-internal object AndroidLibraryConfigurator : ConfigurationContract.AndroidLibraryConfigurator {
+internal object AndroidApplicationConfigurator : ConfigurationContract.AndroidApplicationConfigurator {
     private fun setupAndroidExtension(
-        extension: LibraryExtension,
-        configuration: ConfigurationApiContract.AndroidLibraryConfiguration,
+        extension: ApplicationExtension,
+        configuration: AndroidConfigurationApiContract.AndroidApplicationConfiguration,
     ) {
         extension.compileSdk = configuration.compileSdkVersion
-        extension.resourcePrefix = configuration.prefix
 
         extension.defaultConfig {
             minSdk = configuration.minSdkVersion
@@ -44,33 +43,18 @@ internal object AndroidLibraryConfigurator : ConfigurationContract.AndroidLibrar
             test.java.setSrcDirs(configuration.unitTestSource.sourceDirectories)
             test.res.setSrcDirs(configuration.unitTestSource.resourceDirectories)
 
-            if (configuration.androidTest is ConfigurationApiContract.TestSource) {
-                val androidTest = getByName("androidTest")
-                androidTest.java.setSrcDirs(configuration.androidTest!!.sourceDirectories)
-                androidTest.res.setSrcDirs(configuration.androidTest!!.resourceDirectories)
-            }
-        }
-    }
-
-    private fun setupKmp(
-        project: Project,
-        configuration: ConfigurationApiContract.AndroidLibraryConfiguration,
-    ) {
-        project.extensions.configure(KotlinMultiplatformExtension::class.java) {
-            android().publishLibraryVariants(*configuration.publishVariants.toTypedArray())
+            val androidTest = getByName("androidTest")
+            androidTest.java.setSrcDirs(configuration.androidTest.sourceDirectories)
+            androidTest.res.setSrcDirs(configuration.androidTest.resourceDirectories)
         }
     }
 
     override fun configure(
         project: Project,
-        configuration: ConfigurationApiContract.AndroidLibraryConfiguration,
+        configuration: AndroidConfigurationApiContract.AndroidApplicationConfiguration,
     ) {
-        project.extensions.configure(LibraryExtension::class.java) {
+        project.extensions.configure(ApplicationExtension::class.java) {
             setupAndroidExtension(this, configuration)
-        }
-
-        if (project.isKmp()) {
-            setupKmp(project, configuration)
         }
     }
 }
