+++ antibytes-configuration/src/main/kotlin/tech/antibytes/gradle/configuration/api/AndroidLibraryConfiguration.kt
@@ -6,18 +6,18 @@
 
 package tech.antibytes.gradle.configuration.api
 
-import tech.antibytes.gradle.configuration.ConfigurationApiContract
+import tech.antibytes.gradle.configuration.AndroidConfigurationApiContract
 
 internal data class AndroidLibraryConfiguration(
     override val compileSdkVersion: Int,
     override val minSdkVersion: Int,
     override val targetSdkVersion: Int,
     override val prefix: String,
-    override val compatibilityTargets: ConfigurationApiContract.Compatibility,
+    override val compatibilityTargets: AndroidConfigurationApiContract.Compatibility,
     override val fallbacks: Map<String, Set<String>>,
-    override val mainSource: ConfigurationApiContract.MainSource,
-    override val unitTestSource: ConfigurationApiContract.TestSource,
+    override val mainSource: AndroidConfigurationApiContract.MainSource,
+    override val unitTestSource: AndroidConfigurationApiContract.TestSource,
     override val publishVariants: Set<String>,
-    override val androidTest: ConfigurationApiContract.TestSource?,
-    override val testRunner: ConfigurationApiContract.TestRunner,
-) : ConfigurationApiContract.AndroidLibraryConfiguration
+    override val androidTest: AndroidConfigurationApiContract.TestSource?,
+    override val testRunner: AndroidConfigurationApiContract.TestRunner,
+) : AndroidConfigurationApiContract.AndroidLibraryConfiguration
