+++ antibytes-configuration/src/main/kotlin/tech/antibytes/gradle/configuration/api/TestSource.kt
@@ -6,9 +6,9 @@
 
 package tech.antibytes.gradle.configuration.api
 
-import tech.antibytes.gradle.configuration.ConfigurationApiContract
+import tech.antibytes.gradle.configuration.AndroidConfigurationApiContract
 
 internal data class TestSource(
     override val sourceDirectories: Set<String>,
     override val resourceDirectories: Set<String>,
-) : ConfigurationApiContract.TestSource
+) : AndroidConfigurationApiContract.TestSource
