+++ antibytes-configuration/src/test/kotlin/tech/antibytes/gradle/configuration/AndroidLibraryConfiguratorSpec.kt
@@ -26,6 +26,7 @@ import org.gradle.api.plugins.ExtensionContainer
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinAndroidTarget
 import org.junit.jupiter.api.Test
+import tech.antibytes.gradle.configuration.android.AndroidLibraryConfigurator
 import tech.antibytes.gradle.configuration.api.AndroidLibraryConfiguration
 import tech.antibytes.gradle.configuration.api.Compatibility
 import tech.antibytes.gradle.configuration.api.MainSource
