+++ antibytes-configuration/src/test/kotlin/tech/antibytes/gradle/configuration/DefaultIosDeviceResolverSpec.kt
@@ -15,11 +15,11 @@ import io.mockk.unmockkStatic
 import io.mockk.verify
 import org.gradle.api.NamedDomainObjectCollection
 import org.gradle.api.NamedDomainObjectContainer
-import org.jetbrains.kotlin.gradle.targets.native.KotlinNativeSimulatorTestRun
-import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 import org.gradle.internal.os.OperatingSystem
+import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 import org.jetbrains.kotlin.gradle.plugin.KotlinTarget
 import org.jetbrains.kotlin.gradle.plugin.mpp.KotlinNativeTargetWithSimulatorTests
+import org.jetbrains.kotlin.gradle.targets.native.KotlinNativeSimulatorTestRun
 import org.junit.jupiter.api.AfterEach
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
